move some Volume (soon Item) management ops to content app
