--- loc_insurancemaps/management/volume.py
@@ -1,323 +1,0 @@
-import os
-import glob
-import json
-import logging
-from datetime import datetime
-from osgeo import gdal
-
-from cogeo_mosaic.mosaic import MosaicJSON
-from cogeo_mosaic.backends import MosaicBackend
-
-from django.conf import settings
-from django.core.files import File
-
-from georeference.models.resources import Layer
-
-from loc_insurancemaps.models import Volume
-from loc_insurancemaps.utils import (
-    LOCParser,
-    LOCConnection,
-    filter_volumes_for_use,
-    unsanitize_name
-)
-
-from georeference.utils import (
-    random_alnum,
-)
-
-logger = logging.getLogger(__name__)
-
-def import_all_available_volumes(state, apply_filter=True, verbose=False):
-    """Preparatory step that runs through all cities in the provided
-    state, filters the available volumes for those cities, and then
-    imports each one to create a new Volume object."""
-
-    lc = LOCConnection(delay=0, verbose=verbose)
-    cities = lc.get_city_list_by_state(state)
-
-    volumes = []
-    for city in cities:
-        lc.reset()
-        city = unsanitize_name(state, city[0])
-        vols = lc.get_volume_list_by_city(city, state)
-        if apply_filter is True:
-            vols = filter_volumes_for_use(vols)
-            volumes += [i for i in vols if i['include'] is True]
-        else:
-            volumes += vols
-
-    for volume in volumes:
-        try:
-            Volume.objects.get(pk=volume['identifier'])
-        except Volume.DoesNotExist:
-            import_volume(volume['identifier'])
-
-def import_volume(identifier, locale=None):
-
-    try:
-        volume = Volume.objects.get(pk=identifier)
-        volume.locales.add(locale)
-        volume.update_place_counts()
-        return volume
-    except Volume.DoesNotExist:
-        pass
-
-    lc = LOCConnection(delay=0, verbose=True)
-    response = lc.get_item(identifier)
-    if response.get("status") == 404:
-        return None
-
-    parsed = LOCParser(item=response['item'])
-    volume_kwargs = parsed.volume_kwargs()
-
-    # add resources to args, not in item (they exist adjacent)
-    volume_kwargs["lc_resources"] = response['resources']
-
-    volume = Volume.objects.create(**volume_kwargs)
-    volume.locales.add(locale)
-
-    volume.update_place_counts()
-
-    return volume
-
-def generate_mosaic_geotiff(identifier):
-    """ A helpful reference from the BPLv used during the creation of this method:
-    https://github.com/bplmaps/atlascope-utilities/blob/master/new-workflow/atlas-tools.py
-    """
-
-    start = datetime.now()
-
-    vol = Volume.objects.get(pk=identifier)
-    multimask_geojson = vol.get_multimask_geojson()
-    multimask_file_name = f"multimask-{identifier}"
-    multimask_file = os.path.join(settings.TEMP_DIR, f"{multimask_file_name}.geojson")
-    with open(multimask_file, "w") as out:
-        json.dump(multimask_geojson, out, indent=1)
-
-    trim_list = []
-    for feature in multimask_geojson['features']:
-
-        layer_name = feature['properties']['layer']
-        wo = gdal.WarpOptions(
-            format="VRT",
-            dstSRS = "EPSG:3857",
-            cutlineDSName = multimask_file,
-            cutlineLayer = multimask_file_name,
-            cutlineWhere = f"layer='{layer_name}'",
-            cropToCutline = True,
-            # creationOptions = ['COMPRESS=LZW', 'BIGTIFF=YES'],
-            # resampleAlg = 'cubic',
-            # dstAlpha = False,
-            dstNodata = "255 255 255",
-        )
-
-        layer = Layer.objects.get(slug=layer_name)
-        if not layer.file:
-            raise Exception(f"no layer file for this layer {layer_name}")
-        in_path = layer.file.path
-        trim_name = os.path.basename(in_path).replace(".tif", "_trim.vrt")
-        out_path = os.path.join(settings.TEMP_DIR, trim_name)
-        gdal.Warp(out_path, in_path, options=wo)
-
-        to = gdal.TranslateOptions(
-            format="VRT",
-            bandList = [1,2,3]
-        )
-
-        noalpha_path = out_path.replace(".vrt", "_noalpha.vrt")
-        gdal.Translate(noalpha_path, out_path, options=to)
-
-        trim_list.append(noalpha_path)
-
-    mosaic_vrt = os.path.join(settings.TEMP_DIR, f"{identifier}.vrt")
-    bounds = vol.extent.transform(3857, True).extent
-    vo = gdal.BuildVRTOptions(
-        resolution = 'highest',
-        outputSRS="EPSG:3857",
-        outputBounds=bounds,
-        separate = False,
-        srcNodata = "255 255 255",
-    )
-    print("building vrt")
-
-    use_vrt = False
-    if use_vrt:
-        mosaic_vrt = os.path.join("/opt/app/uploaded/mosaics", f"{identifier}.vrt")
-        gdal.BuildVRT(mosaic_vrt, trim_list, options=vo)
-        print("saving VRT to volume instance")
-        with open(mosaic_vrt, 'rb') as f:
-            vol.mosaic_geotiff = File(f, name=os.path.basename(mosaic_vrt))
-            vol.save()
-        return
-
-    mosaic_vrt = os.path.join(settings.TEMP_DIR, f"{identifier}.vrt")
-    gdal.BuildVRT(mosaic_vrt, trim_list, options=vo)
-
-    print("building final geotiff")
-
-    to = gdal.TranslateOptions(
-        # format="GTiff",
-        format="COG",
-        creationOptions = [
-            # for the GTiff format
-            # "TILED=YES",
-            # "COMPRESS=DEFLATE",
-            # "PREDICTOR=2",
-            # "NUM_THREADS=ALL_CPUS",
-            ## the following is apparently in the COG spec but doesn't work??
-            # "COPY_SOURCE_OVERVIEWS=YES",
-            # for the COG format
-            "COMPRESS=JPEG",
-        ],
-    )
-
-    mosaic_tif = mosaic_vrt.replace(".vrt", ".tif")
-    gdal.Translate(mosaic_tif, mosaic_vrt, options=to)
-
-    ## for some reason, creating overviews and then saving the file to
-    ## django's filefield with
-    ## with open(mosaic_tif, 'rb') as f:
-    ##    vol.mosaic_geotiff.save(os.path.basename(mosaic_tif), File(f), save=True)
-    ## seems to make an empty tiff... (still need to figure out why...)
-    ## for now, save to django without overviews.
-    ## however, Titiler is still returning empty tiles...
-
-    existing_file_path = None
-    if vol.mosaic_geotiff:
-        existing_file_path = vol.mosaic_geotiff.path
-
-    file_name = f"{identifier}__{random_alnum(6)}.tif"
-
-    with open(mosaic_tif, 'rb') as f:
-        vol.mosaic_geotiff.save(file_name, File(f))
-
-    # print("creating overviews")
-    # img = gdal.Open(vol.mosaic_geotiff.path, 1)
-    # gdal.SetConfigOption("COMPRESS_OVERVIEW", "DEFLATE")
-    # gdal.SetConfigOption("PREDICTOR", "2")
-    # gdal.SetConfigOption("GDAL_NUM_THREADS", "ALL_CPUS")
-    # img.BuildOverviews("AVERAGE", [2, 4, 8, 16])
-
-    os.remove(mosaic_tif)
-    if existing_file_path:
-        os.remove(existing_file_path)
-
-    print(f"completed - elapsed time: {datetime.now() - start}")
-
-
-def write_trim_feature_cache(feature, file_path):
-    with open(file_path, "w") as f:
-        json.dump(feature, f, indent=2)
-
-def read_trim_feature_cache(file_path):
-    with open(file_path, "r") as f:
-        feature = json.load(f)
-    return feature
-
-def generate_mosaic_json(identifier, trim_all=False):
-
-    logger.info(f"{identifier} | generating mosaic json")
-    vol = Volume.objects.get(pk=identifier)
-    multimask_geojson = vol.get_multimask_geojson()
-    multimask_file_name = f"multimask-{identifier}"
-    multimask_file = os.path.join(settings.TEMP_DIR, f"{multimask_file_name}.geojson")
-    with open(multimask_file, "w") as out:
-        json.dump(multimask_geojson, out, indent=1)
-
-    logger.debug(f"{identifier} | multimask loaded")
-    logger.info(f"{identifier} | iterating and trimming layers")
-    trim_list = []
-    for feature in multimask_geojson['features']:
-
-        layer_name = feature['properties']['layer']
-        layer = Layer.objects.get(slug=layer_name)
-        if not layer.file:
-            logger.error(f"{identifier} | no layer file for this layer {layer_name}")
-            raise Exception(f"no layer file for this layer {layer_name}")
-        in_path = layer.file.path
-
-        layer_dir = os.path.dirname(in_path)
-        file_name = os.path.basename(in_path)
-        logger.debug(f"{identifier} | processing layer file {file_name}")
-
-        file_root = os.path.splitext(file_name)[0]
-        existing_trimmed_tif = glob.glob(f"{layer_dir}/{file_root}*_trim.tif")
-        print(existing_trimmed_tif)
-
-        feat_cache_path = in_path.replace(".tif", "_trim-feature.json")
-        if os.path.isfile(feat_cache_path):
-            cached_feature = read_trim_feature_cache(feat_cache_path)
-            logger.debug(f"{identifier} | using cached trim json boundary")
-        else:
-            cached_feature = None
-            write_trim_feature_cache(feature, feat_cache_path)
-
-        unique_id = random_alnum(6)
-        trim_vrt_path = in_path.replace(".tif", f"_{unique_id}_trim.vrt")
-        out_path = trim_vrt_path.replace(".vrt", ".tif")
-
-        # compare this multimask feature to the cached one for this layer
-        # and only (re)create a trimmed tif if they do not match
-        if feature != cached_feature or trim_all is True:
-
-            wo = gdal.WarpOptions(
-                format="VRT",
-                dstSRS = "EPSG:3857",
-                cutlineDSName = multimask_file,
-                cutlineLayer = multimask_file_name,
-                cutlineWhere = f"layer='{layer_name}'",
-                cropToCutline = True,
-                creationOptions = ['COMPRESS=LZW', 'BIGTIFF=YES'],
-                resampleAlg = 'cubic',
-                dstAlpha = False,
-                dstNodata = "255 255 255",
-            )
-            gdal.Warp(trim_vrt_path, in_path, options=wo)
-
-            to = gdal.TranslateOptions(
-                format="GTiff",
-                bandList = [1,2,3],
-                creationOptions = [
-                    "TILED=YES",
-                    "COMPRESS=LZW",
-                    "PREDICTOR=2",
-                    "NUM_THREADS=ALL_CPUS",
-                    ## the following is apparently in the COG spec but doesn't work??
-                    # "COPY_SOURCE_OVERVIEWS=YES",
-                ],
-            )
-
-            logger.debug(f"writing trimmed tif {os.path.basename(out_path)}")
-            result = gdal.Translate(out_path, trim_vrt_path, options=to)
-            write_trim_feature_cache(feature, feat_cache_path)
-
-            img = gdal.Open(out_path, 1)
-            if img is None:
-                logger.warn(f"{identifier} | file was not properly created, omitting: {file_name}")
-                continue   
-            logger.debug(f"{identifier} | building overview: {file_name}")
-            gdal.SetConfigOption("COMPRESS_OVERVIEW", "LZW")
-            gdal.SetConfigOption("PREDICTOR", "2")
-            gdal.SetConfigOption("GDAL_NUM_THREADS", "ALL_CPUS")
-            img.BuildOverviews("AVERAGE", [2, 4, 8, 16])
-
-        else:
-            logger.debug(f"{identifier} | using existing trimmed tif {file_name}")
-
-        trim_list.append(out_path)
-
-    trim_urls = [
-        i.replace(os.path.dirname(settings.MEDIA_ROOT), settings.MEDIA_HOST.rstrip("/")) \
-            for i in trim_list
-    ]
-    logger.info(f"{identifier} | writing mosaic from {len(trim_urls)} trimmed tifs")
-    mosaic_data = MosaicJSON.from_urls(trim_urls, minzoom=14)
-    mosaic_json_path = os.path.join(settings.TEMP_DIR, f"{identifier}-mosaic.json")
-    with MosaicBackend(mosaic_json_path, mosaic_def=mosaic_data) as mosaic:
-        mosaic.write(overwrite=True)
-
-    with open(mosaic_json_path, 'rb') as f:
-        vol.mosaic_json = File(f, name=os.path.basename(mosaic_json_path))
-        vol.save()
-
-    logger.info(f"{identifier} | mosaic created: {os.path.basename(mosaic_json_path)}")
